Objects added: Ore, Anvil, furnace, bank booth
Added one of each ore type (except Rune Ess due to it not working) to
the north of the Catherby (non-combat skill training) teleport area. As
well, there is an anvil and furnace for smelting and forging (which
currently produces coins for some reason, no matter what you smith) and
a bank booth, for quick access to your bank while mining.

The changes I made are found in client.java, and can be located around
line 3576, which is where the new stuff I added begins.

I plan to add a woodcutting area as well soon.